+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for SD-WAN Manager 20.15
 - In add task, added comment that Validator takes few minutes to build the control connections
+- In add task, when requested image version is not available, added print of available images
 
 # Catalyst SD-WAN Lab 2.0.12 [Jun 18, 2024]